Use rails_sql_views to generate database migration
This ensures the view migration is dumped in the schema.rb. Note the original
migration was overwritten because this should have no functional effect on the
database structure, apart from now being dumped to the schema.rb file.

The dummy app's db/migrate is symlinked to the gem's db/migrate to keep the
timestamps in sync.